Add assos seeder

+++ database/migrations/2010_10_12_000000_clear.php
@@ -10,7 +10,7 @@ class Clear extends Migration
      */
     public function up()
     {
-        $this->removeDir(public_path('/images/assos'));
+        $this->removeDir(public_path('/images/assos/'), true);
         $this->removeDir(public_path('/images/articles'));
         $this->removeDir(public_path('/images/events'));
     }
@@ -18,22 +18,25 @@ class Clear extends Migration
     /**
      * Delete a folder and all its content.
      *
-     * @param  string $path
+     * @param string  $path
+     * @param boolean $contentOnly
      * @return void
      */
-    protected function removeDir(string $path): void
+    protected function removeDir(string $path, bool $contentOnly=false): void
     {
         if (file_exists($path)) {
-            foreach (array_diff(scandir($path), ['..', '.']) as $file) {
+            foreach (array_diff(scandir($path), ['..', '.', '.gitignore']) as $file) {
                 $subPath = $path.DIRECTORY_SEPARATOR.$file;
                 if (is_dir($subPath)) {
-                    $this->removeDir($subPath);
+                    $this->removeDir($subPath, $contentOnly);
                 } else {
                     unlink($subPath);
                 }
             }
 
-            rmdir($path);
+            if (!$contentOnly) {
+                rmdir($path);
+            }
         }
     }
